Extracted model view to seperate class
