+++ compliance-by-design-demo/package.json
@@ -5,7 +5,7 @@
   "homepage": ".",
   "dependencies": {
     "@discipl/core-ephemeral": "0.10.1",
-    "@discipl/law-reg": "^0.2.0",
+    "@discipl/law-reg": "^0.2.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "react-scripts": "3.0.1"
