+++ compliance-by-design-demo/src/components/ActorView.js
