+++ compliance-by-design-demo/src/model/lerarenbeurs.json
@@ -181,13 +181,43 @@
             "sourcetext": ""
         },
         {
+            "explanation": "Added by Pim",
+            "act": "<<minister verzoekt bewijs van betaling van collegegeld>>",
+            "actor": "[minister van Onderwijs, Cultuur en Wetenschap]",
+            "action": "[verzoeken]",
+            "object": "[verzoek tot bewijs van betaling van collegegeld]",
+            "interested-party": "[leraar]",
+            "preconditions": "[besluit tot verlenen subsidie voor studiekosten van een leraar in verband met het volgen van een opleiding]",
+            "create": "[verzoek tot bewijs van betaling van collegegeld]",
+            "terminate": "",
+            "version": "",
+            "reference": "art. 19 Slb",
+            "juriconnect": "jci1.3:c:BWBR0039319&hoofdstuk=2&artikel=19&z=2019-04-01&g=2019-04-01",
+            "sourcetext": ""
+        },
+        {
+            "explanation": "Added by Pim",
+            "act": "<<minister verzoekt bewijs van het behalen van ten minste vijftien studiepunten>>",
+            "actor": "[minister van Onderwijs, Cultuur en Wetenschap]",
+            "action": "[verzoeken]",
+            "object": "[verzoek tot bewijs van het behalen van ten minste vijftien studiepunten]",
+            "interested-party": "[leraar]",
+            "preconditions": "[besluit tot verlenen subsidie voor studiekosten van een leraar in verband met het volgen van een opleiding]",
+            "create": "[verzoek tot bewijs van het behalen van ten minste vijftien studiepunten]",
+            "terminate": "",
+            "version": "",
+            "reference": "art. 19 Slb",
+            "juriconnect": "jci1.3:c:BWBR0039319&hoofdstuk=2&artikel=19&z=2019-04-01&g=2019-04-01",
+            "sourcetext": ""
+        },
+        {
             "explanation": "",
             "act": "<<leraar overlegt bewijsstuk waaruit blijkt dat hij ten minste vijftien studiepunten heeft gehaald>>",
             "actor": "[leraar]",
             "action": "[overleggen]",
             "object": "[bewijsstuk waaruit blijkt dat hij ten minste vijftien studiepunten heeft gehaald]",
             "interested-party": "[minister van Onderwijs, Cultuur en Wetenschap]",
-            "preconditions": "[subsidiebedrag is uitbetaald aan subsidieontvanger voordat de opleiding waar de subsidie betrekking op heeft aanvangt]\r\nEN\r\n[studiejaar waarop subsidie betrekking heeft is voorbij]",
+            "preconditions": "[verzoek tot bewijs van het behalen van ten minste vijftien studiepunten] EN [subsidiebedrag is uitbetaald aan subsidieontvanger voordat de opleiding waar de subsidie betrekking op heeft aanvangt]\r\nEN\r\n[studiejaar waarop subsidie betrekking heeft is voorbij]",
             "create": "",
             "terminate": "",
             "version": "",
@@ -202,7 +232,7 @@
             "action": "[overleggen]",
             "object": "[bewijsstuk waaruit blijkt dat hij collegegeld heeft betaald]",
             "interested-party": "[minister van Onderwijs, Cultuur en Wetenschap]",
-            "preconditions": "[subsidiebedrag is uitbetaald aan subsidieontvanger voordat de opleiding waar de subsidie betrekking op heeft aanvangt]\r\nEN\r\n[studiejaar waarop subsidie betrekking heeft is voorbij]",
+            "preconditions": "[verzoek tot bewijs van betaling van collegegeld] EN [subsidiebedrag is uitbetaald aan subsidieontvanger voordat de opleiding waar de subsidie betrekking op heeft aanvangt]\r\nEN\r\n[studiejaar waarop subsidie betrekking heeft is voorbij]",
             "create": "",
             "terminate": "",
             "version": "",
@@ -322,7 +352,7 @@
             "action": "[weigeren]",
             "object": "[subsidieverlening aan een leraar]",
             "interested-party": "[leraar]",
-            "preconditions": "[besluit tot verlenen subsidie voor kosten in verband met het verlenen van studieverlof aan de leraar]",
+            "preconditions": "[besluit tot verlenen subsidie voor studiekosten van een leraar in verband met het volgen van een opleiding]",
             "create": "",
             "terminate": "[aanvraag]",
             "version": "",
@@ -914,6 +944,24 @@
         {
             "explanation": "",
             "fact": "[besluit tot verlenen subsidie voor studiekosten van een leraar in verband met het volgen van een opleiding]",
+            "function": "<<>>",
+            "reference": "",
+            "version": "",
+            "juriconnect": "",
+            "sourcetext": "Added, because it was missing in excel"
+        },
+        {
+            "explanation": "",
+            "fact": "[verzoek tot bewijs van het behalen van ten minste vijftien studiepunten]",
+            "function": "<<>>",
+            "reference": "",
+            "version": "",
+            "juriconnect": "",
+            "sourcetext": "Added, because it was missing in excel"
+        },
+        {
+            "explanation": "",
+            "fact": "[verzoek tot bewijs van betaling van collegegeld]",
             "function": "<<>>",
             "reference": "",
             "version": "",
